Moving Hooks into directory
